+++ app/models/dmsf/file.rb
@@ -1,0 +1,5 @@
+module Dmsf
+  class File < Entity
+
+  end
+end