+++ app/models/dmsf/folder.rb
@@ -1,0 +1,5 @@
+module Dmsf
+  class Folder < Entity
+
+  end
+end