+++ test/unit/dmsf_file_test.rb
@@ -1,8 +1,21 @@
-require "test/unit"
+require File.dirname(__FILE__) + '/../test_helper'
 
-class DmsfFileTest < DmsfEntityTest
+class DmsfFileTest < Test::UnitTest
+
+
+  #We need to check a link the model creates to a non-existant user and project
+  fixtures :projects, :users
 
   context "Object declaration" do
+
+    should "Utilise awesome_nested_set" do
+      assert Dmsf::File.include?(CollectiveIdea::Acts::NestedSet::Model)
+    end
+
+    should "Utilise dmsf_entity database table" do
+      assert_equal Dmsf::Entity.table_name, Dmsf::File.table_name
+    end
+
     should "extend Dmsf::Entity" do
       Dmsf::File.superclass.is_a?(Dmsf::Entity)
     end